+++ gpaw/calculator.py
@@ -29,7 +29,6 @@ from gpaw.io.logger import GPAWLogger
 from gpaw.jellium import create_background_charge
 from gpaw.kohnsham_layouts import get_KohnSham_layouts
 from gpaw.kpt_descriptor import KPointDescriptor
-from gpaw.kpt_refine import create_kpoint_descriptor_with_refinement
 from gpaw.matrix import suggest_blocking
 from gpaw.occupations import ParallelLayout, create_occ_calc
 from gpaw.output import (print_cell, print_parallelization_details,
@@ -1264,26 +1263,11 @@ class GPAW(Calculator):
                       pbc=self.atoms.pbc)
         bzkpts_kc = kpts2ndarray(par.kpts, atoms)
 
-        kpt_refine = par.experimental.get('kpt_refine')
-        if kpt_refine is None:
-            kd = KPointDescriptor(bzkpts_kc, nspins)
+        kd = KPointDescriptor(bzkpts_kc, nspins)
 
-            self.timer.start('Set symmetry')
-            kd.set_symmetry(self.atoms, self.symmetry, comm=self.world)
-            self.timer.stop('Set symmetry')
-
-        else:
-            self.timer.start('Set k-point refinement')
-            kd = create_kpoint_descriptor_with_refinement(
-                kpt_refine,
-                bzkpts_kc, nspins, self.atoms,
-                self.symmetry, comm=self.world,
-                timer=self.timer)
-            self.timer.stop('Set k-point refinement')
-            # Update quantities which might have changed, if symmetry
-            # was changed
-            self.symmetry = kd.symmetry
-            self.setups.set_symmetry(kd.symmetry)
+        self.timer.start('Set symmetry')
+        kd.set_symmetry(self.atoms, self.symmetry, comm=self.world)
+        self.timer.stop('Set symmetry')
 
         self.log(kd)